+++ conformance/config.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"path/filepath"
@@ -19,11 +20,6 @@ const (
 	// directory containing all the cwl files and input json/yamls
 	// github.com/uc-cdis/mariner/conformance/common-workflow-language/v1.0/v1.0
 	pathToTests = "./common-workflow-language/v1.0/v1.0/"
-
-	// all path/location of files need this prefix
-	// in the user data s3 bucket, the directory structure is:
-	// -- /userID/conformanceTesting/<file>
-	inputPathPrefix = "USER/conformanceTesting/"
 )
 
 // TestCase ..
@@ -36,6 +32,12 @@ type TestCase struct {
 	ID         int                    `json:"id" yaml:"id"`
 	Doc        string                 `json:"doc" yaml:"doc"`
 	Tags       []string               `json:"tags" yaml:"tags"`
+}
+
+// Creds is creds.json, as downloaded from the portal
+type Creds struct {
+	APIKey string `json:"api_key"`
+	KeyID  string `json:"key_id"`
 }
 
 func loadConfig(config string) ([]*TestCase, error) {
@@ -57,4 +59,26 @@ func loadConfig(config string) ([]*TestCase, error) {
 	}
 
 	return *testSuite, nil
+}
+
+// read creds into Creds struct
+func apiKey(creds string) ([]byte, error) {
+	// read in bytes
+	b, err := ioutil.ReadFile(creds)
+	if err != nil {
+		return nil, err
+	}
+
+	// validate against creds schema
+	c := &Creds{}
+	err = json.Unmarshal(b, c)
+	if err != nil {
+		return nil, err
+	}
+	if c.APIKey == "" || c.KeyID == "" {
+		return nil, fmt.Errorf("missing credentials")
+	}
+
+	// return bytes
+	return b, nil
 }
